Fixed bug in the loop that finds keywords in articles
The loop that searched for keywords in articles and saved the text and
date in separate lists should have a list of their titles as a control,
so as not to save the same article more than once if the title was
already in this list. However, the titles of the articles found were not
being added to that list.