Add script to remove duplicate articles
It checks for duplicate articles in the 'articles without stop words'
list and creates a new list with its index. The user could use this list
to check what happened to those articles.

It removes duplicate articles and their date of publication from the
lists.

It also tells the user how many items were on the list before and after
this process.