Add script to build the general NLP model
It uses Word2Vec to make a model with all the articles present in the
list of unique articles.

It informs the user how many words are part of the vocabulary and the 10
most frequent words.

Finally, it finds 'most similar words' to a word of interest, in this
case, 'quarantine'.